+++ src/storage/mvcc/reader/mod.rs
@@ -5,6 +5,6 @@ mod reader;
 mod scanner;
 
 pub use self::point_getter::{PointGetter, PointGetterBuilder};
-pub use self::reader::{check_need_gc, MvccReader};
+pub use self::reader::{check_need_gc, check_region_need_gc, MvccReader};
 pub use self::scanner::test_util;
 pub use self::scanner::{has_data_in_range, DeltaScanner, EntryScanner, Scanner, ScannerBuilder};
